requirements.txt
Flask>=2.0
Pillow>=9.0
numpy>=1.22
numpy-stl>=2.16